Revert timestamp as a base type
Previous edits extracted the "timestamp" type in custom_types.yml and
set it back as a `uinteger`. However we need at least in Elixir to have
the _timestamp_ type as a basic type, like string or uinteger, in order
to know when to (un)serialize it to the native `DateTime` struct.

This commit removes the "types.timestamp" custom type and changes
wherever it appeared to the basic "timestamp" type.